send remaining pipeline-specific file to assets
